ai_worker/requirements.txt
edge-tts==7.2.7
aiohttp==3.13.2
tabulate==0.9.0
certifi==2025.11.12
requests==2.32.3
openai==1.57.0
python-dotenv==1.0.1
numpy
opencv-python
Pillow
torch
torchaudio
chatterbox-tts
scipy
librosa
transformers
diffusers
safetensors
omegaconf
resemble-perth
conformer
s3tokenizer
spacy-pkuseg
setuptools
# FFmpeg is required but is a system dependency, not a python package.
# These are the core packages needed for the AI worker to function.
